Use intptr_t for all array size parameters (64-bit safe)
All C function signatures changed: int ng/nv/dim/n/npx/ns/nf/N/nnz/nhist/np
→ intptr_t to handle arrays larger than 2^31 elements. Py_ssize_t (64-bit
on 64-bit Python) now maps to intptr_t without truncation.

- .c2py files: int → intptr_t in all c_overloads.sig size params
- C sources: closest.c, cdiffraction.c, darkflat.c, blobs.c,
  connectedpixels.c, localmaxlabel.c, sparse_image.c, splat.c
- SIMD kernels: all 16 kernel files in geometry/simd/ and imageproc/simd/
- _wrappers.c: all forward declarations and bridge functions
- bslz4: bs_functions.h, bs_master.c (via templates), .c2py bslz4 file
- c2py_runtime.c/h: synced from c2py23 (Windows support)
- test.yml: fix diagnostic step pwsh syntax

Thresholds, flags, labels, return types kept as int (always small).

+++ src/geometry/cdiffraction.h
@@ -1,3 +1,4 @@
+#include <stdint.h>
 #define PI 3.14159265358979323846
 #define RAD PI / 180.0
 #define DEG 180.0 / PI
@@ -29,15 +30,15 @@
     (c)[8] = (a)[2] * (b)[6] + (a)[5] * (b)[7] + (a)[8] * (b)[8];
 
 void assign(double ubi[9], double gv[][3], double tol, double drlv2[],
-            int labels[], int ig, int n);
+            int labels[], int ig, intptr_t n);
 
 void compute_gv(double xlylzl[][3], double omega[], double omegasign,
                 double wvln, double wedge, double chi, double t[3],
-                double gv[][3], int n);
+                double gv[][3], intptr_t n);
 
 void compute_xlylzl(double s[], double f[], double p[4], double r[9],
-                    double dist[3], double xlylzl[][3], int n);
+                    double dist[3], double xlylzl[][3], intptr_t n);
 
 void compute_xlylzl_xpos_variable(double s[], double f[], double p[4], double r[9],
     double dist[3], double xpos[],
-    double xlylzl[][3], int n);
+    double xlylzl[][3], intptr_t n);
