+++ src/imageproc/simd/reorder_u16_a32_kernel.c
@@ -8,11 +8,12 @@
 
 #include "cImageD11.h"
 #include <stdio.h>
+#include <stdint.h>
 
 void KERNEL_FN(const uint16_t *restrict data,
                      const uint32_t *restrict adr, uint16_t *restrict out,
-                     int N) {
-    int i;
+                     intptr_t N) {
+    intptr_t i;
     /*  printf("Hello, got N=%d\n",N);*/
 #pragma omp parallel for
     for (i = 0; i < N; i++) {
