+++ src/imageproc/simd/reorderlut_u16_a32_kernel.c
@@ -8,10 +8,11 @@
 
 #include "cImageD11.h"
 #include <stdio.h>
+#include <stdint.h>
 
 void KERNEL_FN(uint16_t *restrict data, uint32_t *restrict lut,
-                        uint16_t *restrict out, int N) {
-    int i;
+                        uint16_t *restrict out, intptr_t N) {
+    intptr_t i;
     /*  printf("Hello, got N=%d\n",N);*/
 #pragma omp parallel for
     for (i = 0; i < N; i++) {
